Consolidate response body specializations as inner classes.
